requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
filelock @ file:///croot/filelock_1700591183607/work
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
huggingface-hub==0.24.6
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
lightning-utilities==0.11.6
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
openai-clip==1.0.1
packaging==24.1
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.4
six==1.16.0
sympy @ file:///croot/sympy_1701397643339/work
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.1
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth==0.2.13
yarl==1.9.4
